Finalized Catalog Item Component, logic running as usual.
